Converting project/solution to Visual Studio 2015:   Migrating input layer to DirectInput 8 (v7 is no longer supported)   Fixing various warnings up to /W level 3.     (Higher levels have problems with libsnss, nes_apu.c, and other code that      is kinda considered 'external' to nester. Probably easy to fix, but since      snss involves loading of NES save-state files, testing would be needed)   Removing old postbuild events. UPX just tends to trigger virus scanners nowadays.

+++ src/nes/cpu/nes_6502.cpp
@@ -40,14 +40,14 @@ static nes6502_memread NESReadHandler[] =
 {
    /* $0 - $7FF is RAM */
    { 0x0800, 0xFFFF, NES_read },
-   { -1,     -1,     NULL }
+   { (uint32)-1, (uint32)-1, NULL }
 };
 
 static nes6502_memwrite NESWriteHandler[] =
 {
    /* $0 - $7FF is RAM */
    { 0x0800, 0xFFFF, NES_write },
-   { -1,     -1,     NULL}
+   { (uint32)-1, (uint32)-1, NULL}
 };
